pyproject.toml
[build-system]
requires = ["uv_build"]
build-backend = "uv_build"

[project]
name = "booster-rpc"
version = "0.0.8"
description = "Python client for controlling Booster K1 humanoid robots via gRPC and WebSocket"
authors = [
  { name = "Miguel Villa Floran", email = "miguel.villafloran@gmail.com" },
]
requires-python = ">=3.10,<3.14"
readme = "README.md"
license = { file = "LICENSE" }
keywords = ["robotics", "robot", "control", "humanoid", "booster", "k1"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

dependencies = [
  "betterproto>=1.2.5",
  "grpcio>=1.78.1",
  "numpy>=1.26.0",
  "opencv-python>=4.13.0.92",
  "websockets>=16.0",
]

[dependency-groups]
dev = ["ipykernel", "pre-commit"]
